requirements.txt
asgiref==3.5.2
colorama==0.4.5
Django==4.1.1
Pillow==9.2.0
psycopg2==2.9.3
qrcode==7.3.1
sqlparse==0.4.3
tzdata==2022.4
